Move wagtail.contrib.forms.edit_handlers to wagtail.contrib.forms.panels
