+++ wagtail/contrib/forms/tests/test_views.py
@@ -12,8 +12,8 @@ from openpyxl import load_workbook
 
 from wagtail.admin.forms import WagtailAdminPageForm
 from wagtail.admin.panels import get_form_for_model
-from wagtail.contrib.forms.edit_handlers import FormSubmissionsPanel
 from wagtail.contrib.forms.models import FormSubmission
+from wagtail.contrib.forms.panels import FormSubmissionsPanel
 from wagtail.contrib.forms.tests.utils import (
     make_form_page,
     make_form_page_with_custom_submission,
